Renaming handle -> device_handle and using handle as a typedef for now.
